validate the input data, if none, continue
fix bug

align the version

add method to generate tasks from geojson

fix bug

add geojson to tasks

add unit tests

add notebook for clipping tasks to aez

add command line notes

add notebook for clipping tasks to aez

ensure southern testing notebooks are clean and working

add heart beat as default

use heatbeat branch for pip install

add 900 seconds to sqs invisability

start northern cropmask

use uint8

+++ production/dea_ml/dea_ml/core/predict_from_feature.py
@@ -160,8 +160,8 @@ def predict_with_model(
         return_input=True,
     )
 
-    predicted["Predictions"] = predicted["Predictions"].astype("int8")
-    predicted["Probabilities"] = predicted["Probabilities"].astype("float32")
+    predicted["Predictions"] = predicted["Predictions"].astype("uint8")
+    predicted["Probabilities"] = predicted["Probabilities"].astype("uint8")
 
     return predicted
